Closes #25. Apikey not required anymore
